generate_2fa.py вынесен в отдельный файл Добавлена очистка кеша в стиме для "Чистого" запуска клиента
